fixed HW1 Q1 problem

+++ Excersises-Quera/Class/HW1/HW1_Q1_402170516.cpp
@@ -3,13 +3,13 @@
 using namespace std;
 int main()
 {
-  int n, m, k;
+  long int n, m, k;
   string args;
   getline(cin, args);
   istringstream nmk(args);
   nmk >> n >> m >> k;
-  int quantity_x = (n + k - 1) / k; // Round up division for x
-  int quantity_y = (m + k - 1) / k; // Round up division for y
+  long int quantity_x = (n + k - 1) / k; // Round up division for x
+  long int quantity_y = (m + k - 1) / k; // Round up division for y
   cout << quantity_x * quantity_y << endl;
   return 0;
 }
